Add files via upload
Debugged with chatgpt4 that changes how the script counts the transactions. It increases the limit from 100 to 650, and then adds an additional part of the script to ensure people are not manually adding 0 FRY transactions to increase their reward. This script will determine if the amount of transactions within 24 hours is appropriate for the number of rows an address appears in. If the number of transactions is not appropriate for the # of rows an address is in, the script will just do the max reward that is appropriate for the # of rows.